Implement toast after copying text.
